requirements.txt
# Core dependencies for Hugging Face Space
gradio>=5.22.0
openai-agents>=0.3.0
anthropic>=0.49.0
openai>=1.68.2
pypdf>=5.4.0
python-dotenv>=1.0.1
requests>=2.32.3
httpx>=0.28.1
